+++ src/main/scala/ai/tripl/arc/api/API.scala
@@ -262,6 +262,8 @@ object API {
 
     def after(result: Option[DataFrame], stage: PipelineStage, index: Int, stages: List[PipelineStage])(implicit spark: SparkSession, logger: ai.tripl.arc.util.log.logger.Logger, arcContext: ARCContext)
 
+    def runStage(stage: PipelineStage, index: Int, stages: List[PipelineStage])(implicit spark: SparkSession, logger: ai.tripl.arc.util.log.logger.Logger, arcContext: ARCContext): Boolean = true
+
   }
 
 
@@ -297,7 +299,15 @@ object API {
 
   sealed trait Authentication
   object Authentication {
+    /**
+      In the Amazon case, we support encryption options using IAM access only as in order
+      to provide access to a KMS key that would need to be done via a role and SSE-S3 should work if
+      enable on the bucket as the default. Therefor unless a use case appears for adding encryption
+      options to the access key method we will only support encryption options when using IAM for now.
+     */
     case class AmazonAccessKey(accessKeyID: String, secretAccessKey: String, endpoint: Option[String], ssl: Option[Boolean]) extends Authentication
+    case class AmazonIAM(encryptionType: Option[AmazonS3EncryptionType], keyArn: Option[String], customKey: Option[String]) extends Authentication
+    case object AmazonAnonymous extends Authentication
     case class AzureSharedKey(accountName: String, signature: String) extends Authentication
     case class AzureSharedAccessSignature(accountName: String, container: String, token: String) extends Authentication
     case class AzureDataLakeStorageToken(clientID: String, refreshToken: String) extends Authentication
@@ -305,6 +315,23 @@ object API {
     case class AzureDataLakeStorageGen2OAuth(clientID: String, secret: String, directoryId: String) extends Authentication
     case class GoogleCloudStorageKeyFile(projectID: String, keyFilePath: String) extends Authentication
   }
+
+  sealed trait AmazonS3EncryptionType
+  object AmazonS3EncryptionType {
+    case object SSE_S3 extends AmazonS3EncryptionType // AES256
+    case object SSE_KMS extends AmazonS3EncryptionType
+    case object SSE_C extends AmazonS3EncryptionType // custom
+
+    def fromString(encType: String): Option[AmazonS3EncryptionType] = {
+      Option(encType).map(_.trim.toUpperCase) match {
+        case Some("SSE-S3") => Some(SSE_S3)
+        case Some("SSE-KMS") => Some(SSE_KMS)
+        case Some("SSE-C") => Some(SSE_C)
+        case _ => None
+      }
+    }
+  }
+
 
   case class Watermark(eventTime: String, delayThreshold: String)