+++ src/test/resources/conf/sql/false.sql
@@ -1,0 +1,1 @@
+SELECT FALSE, '{"result": false}'